requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.4
Pillow==11.3.0             # ✅ Versión compatible con Python 3.13 y Windows
django-phonenumber-field==7.3.0
phonenumbers==8.13.25
drf-yasg==1.21.7
python-decouple==3.8
whitenoise==6.6.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
reportlab==4.0.7
openpyxl==3.1.2
# Sin Celery ni Redis - sistema simplificado
factory-boy==3.3.0
faker==20.1.0
pyfcm==1.5.4
firebase-admin==6.4.0

